+++ src/lib.rs
@@ -108,6 +108,7 @@ pub use wapc::WasiParams;
 
 pub type SubjectClaimsPair = (String, Claims<wascap::jwt::Actor>);
 
+use crate::inthost::fetch_oci_bytes;
 use bus::{get_namespace_prefix, MessageBus};
 use crossbeam::Sender;
 #[cfg(feature = "lattice")]
@@ -116,6 +117,8 @@ use crossbeam_channel::Receiver;
 #[cfg(any(feature = "lattice", feature = "manifest"))]
 use inthost::RESTRICTED_LABELS;
 use plugins::PluginManager;
+use std::path::Path;
+use std::str::FromStr;
 use std::{
     collections::HashMap,
     sync::{Arc, RwLock},
@@ -153,8 +156,6 @@ pub struct HostBuilder {
     labels: HashMap<String, String>,
     ns: Option<String>,
     authorizer: Box<dyn Authorizer + 'static>,
-    #[cfg(feature = "lattice")]
-    gantry_client: Option<gantryclient::Client>,
 }
 
 impl HostBuilder {
@@ -163,19 +164,10 @@ impl HostBuilder {
     /// labels, the namespace will be gleaned from the `LATTICE_NAMESPACE` environment variable
     /// (if lattice mode is enabled), and the default authorizer will be set.
     pub fn new() -> HostBuilder {
-        #[cfg(not(feature = "lattice"))]
         let b = HostBuilder {
             labels: inthost::detect_core_host_labels(),
             ns: get_namespace_prefix(),
             authorizer: Box::new(authz::DefaultAuthorizer::new()),
-        };
-
-        #[cfg(feature = "lattice")]
-        let b = HostBuilder {
-            labels: inthost::detect_core_host_labels(),
-            ns: get_namespace_prefix(),
-            authorizer: Box::new(authz::DefaultAuthorizer::new()),
-            gantry_client: None,
         };
 
         b
@@ -217,26 +209,12 @@ impl HostBuilder {
         HostBuilder { labels: hm, ..self }
     }
 
-    /// Sets the gantry client to be used by the host.
-    #[cfg(feature = "lattice")]
-    pub fn with_gantryclient(self, client: gantryclient::Client) -> HostBuilder {
-        HostBuilder {
-            gantry_client: Some(client),
-            ..self
-        }
-    }
-
     /// Converts the transient builder instance into a realized host runtime instance
     pub fn build(self) -> Host {
         #[cfg(not(feature = "lattice"))]
         let h = Host::generate(self.authorizer, self.labels, self.ns.clone());
         #[cfg(feature = "lattice")]
-        let h = Host::generate(
-            self.authorizer,
-            self.labels,
-            self.ns.clone(),
-            self.gantry_client.clone(),
-        );
+        let h = Host::generate(self.authorizer, self.labels, self.ns.clone());
         h
     }
 }
@@ -252,11 +230,12 @@ pub struct Host {
     middlewares: Arc<RwLock<Vec<Box<dyn Middleware>>>>,
     // the key to this field is the subscription subject, and not either a pk or a capid
     terminators: Arc<RwLock<HashMap<String, Sender<bool>>>>,
-    #[cfg(feature = "lattice")]
-    gantry_client: Arc<RwLock<Option<gantryclient::Client>>>,
-    key: KeyPair,
+    pk: String,
+    sk: String,
     authorizer: Arc<RwLock<Box<dyn Authorizer>>>,
     labels: Arc<RwLock<HashMap<String, String>>>,
+    // mapping between OCI registry image references and the associated unique identity (e.g. "Mxxx" and "Vxxx")
+    image_map: Arc<RwLock<HashMap<String, String>>>,
     ns: Option<String>,
 }
 
@@ -264,18 +243,10 @@ impl Host {
     /// Creates a new runtime host using all of the default values. Use the host builder
     /// if you want to provide more customization options
     pub fn new() -> Self {
-        #[cfg(not(feature = "lattice"))]
         let h = Self::generate(
             Box::new(authz::DefaultAuthorizer::new()),
             inthost::detect_core_host_labels(),
             get_namespace_prefix(),
-        );
-        #[cfg(feature = "lattice")]
-        let h = Self::generate(
-            Box::new(authz::DefaultAuthorizer::new()),
-            inthost::detect_core_host_labels(),
-            get_namespace_prefix(),
-            None,
         );
         h
     }
@@ -284,7 +255,6 @@ impl Host {
         authz: Box<dyn Authorizer + 'static>,
         labels: HashMap<String, String>,
         ns: Option<String>,
-        #[cfg(feature = "lattice")] gantry: Option<gantryclient::Client>,
     ) -> Self {
         let key = KeyPair::new_server();
         let claims = Arc::new(RwLock::new(HashMap::new()));
@@ -293,6 +263,7 @@ impl Host {
         let labels = Arc::new(RwLock::new(labels));
         let terminators = Arc::new(RwLock::new(HashMap::new()));
         let authz = Arc::new(RwLock::new(authz));
+        let image_map = Arc::new(RwLock::new(HashMap::new()));
 
         #[cfg(feature = "lattice")]
         let (com_s, com_r): (Sender<ControlCommand>, Receiver<ControlCommand>) =
@@ -309,6 +280,7 @@ impl Host {
             ns.clone(),
             com_s,
             authz.clone(),
+            image_map.clone(),
         ));
 
         #[cfg(not(feature = "lattice"))]
@@ -317,7 +289,6 @@ impl Host {
         #[cfg(feature = "lattice")]
         let _ = bus.publish_event(BusEvent::HostStarted(key.public_key()));
 
-        #[cfg(feature = "lattice")]
         let host = Host {
             terminators: terminators.clone(),
             bus: bus.clone(),
@@ -326,27 +297,15 @@ impl Host {
             bindings,
             caps,
             middlewares: Arc::new(RwLock::new(vec![])),
-            gantry_client: Arc::new(RwLock::new(gantry)),
-            key: key,
+            pk: key.public_key(),
+            sk: key.seed().unwrap(),
             authorizer: authz,
             labels,
             ns,
+            image_map,
         };
-        #[cfg(not(feature = "lattice"))]
-        let host = Host {
-            terminators: terminators.clone(),
-            bus: bus.clone(),
-            claims: claims.clone(),
-            plugins: Arc::new(RwLock::new(PluginManager::default())),
-            bindings,
-            middlewares: Arc::new(RwLock::new(vec![])),
-            caps,
-            key: key,
-            authorizer: authz,
-            labels,
-            ns,
-        };
-        info!("Host ID is {} (v{})", host.key.public_key(), VERSION,);
+
+        info!("Host ID is {} (v{})", key.public_key(), VERSION);
 
         host.ensure_extras().unwrap();
 
@@ -356,10 +315,7 @@ impl Host {
         host
     }
 
-    /// Adds an actor to the host. This will provision resources (such as a handler thread) for the actor. Actors
-    /// will not be able to make use of capability providers unless bindings are added (or existed prior to the actor
-    /// being added to a host, which is possible in `lattice` mode)
-    pub fn add_actor(&self, actor: Actor) -> Result<()> {
+    fn add_actor_imgref(&self, actor: Actor, imgref: Option<String>) -> Result<()> {
         if self
             .claims
             .read()
@@ -385,6 +341,7 @@ impl Host {
             actor.token.claims.clone(),
         );
 
+        let key = KeyPair::from_seed(&self.sk).unwrap();
         let wg = crossbeam_utils::sync::WaitGroup::new();
         // Spin up a new thread that listens to "wasmbus.Mxxxx" calls on the message bus
         spawns::spawn_actor(
@@ -400,8 +357,10 @@ impl Host {
             self.bindings.clone(),
             c.clone(),
             self.terminators.clone(),
-            self.key.clone(),
+            key,
             self.authorizer.clone(),
+            self.image_map.clone(),
+            imgref,
         )?;
         wg.wait();
         if actor.capabilities().contains(&extras::CAPABILITY_ID.into()) {
@@ -418,22 +377,21 @@ impl Host {
         Ok(())
     }
 
-    /// Adds an actor to the host by looking it up in a Gantry repository, downloading
-    /// the signed module bytes, and adding them to the host. The connection to a Gantry repository
-    /// will be facilitated by the Gantry client supplied by a `HostBuilder`. By default, hosts
-    /// do not come configured with Gantry clients, so you will have to configure one to use this function
-    #[cfg(feature = "lattice")]
-    pub fn add_actor_from_gantry(&self, actor: &str, revision: u32) -> Result<()> {
-        if self.gantry_client.read().unwrap().is_none() {
-            return Err(errors::new(errors::ErrorKind::MiscHost(
-                "No gantry client configured for this host".to_string(),
-            )));
-        }
+    /// Adds an actor to the host. This will provision resources (such as a handler thread) for the actor. Actors
+    /// will not be able to make use of capability providers unless bindings are added (or existed prior to the actor
+    /// being added to a host, which is possible in `lattice` mode)
+    pub fn add_actor(&self, actor: Actor) -> Result<()> {
+        self.add_actor_imgref(actor, None)
+    }
 
-        let vec =
-            bus::lattice::actor_bytes_from_gantry(actor, self.gantry_client.clone(), revision)?;
+    /// Adds an actor to the host by attempting to retrieve it from an OCI
+    /// registry. This function takes an image reference as an argument, e.g.
+    /// myregistry.mycloud.io/actor:v1
+    /// If OCI credentials are supplied in environment variables, those will be used.
+    pub fn add_actor_from_registry(&self, image: &str) -> Result<()> {
+        let bytes = inthost::fetch_oci_bytes(image)?;
 
-        self.add_actor(Actor::from_slice(&vec)?)?;
+        self.add_actor_imgref(Actor::from_slice(&bytes)?, Some(image.to_string()))?;
         Ok(())
     }
 
@@ -450,6 +408,7 @@ impl Host {
         let binding = binding.unwrap_or("default");
 
         let wg = crossbeam_utils::sync::WaitGroup::new();
+        let key = KeyPair::from_seed(&self.sk).unwrap();
         // Spins up a new thread subscribed to the "wasmbus.{capid}.{binding}" subject
         spawns::spawn_actor(
             wg.clone(),
@@ -464,8 +423,10 @@ impl Host {
             self.bindings.clone(),
             self.claims.clone(),
             self.terminators.clone(),
-            self.key.clone(),
+            key,
             self.authorizer.clone(),
+            self.image_map.clone(),
+            None,
         )?;
         wg.wait();
         Ok(())
@@ -489,7 +450,8 @@ impl Host {
     /// so make sure the new actor can handle this stream of these delayed messages. Also ensure that
     /// the underlying WebAssembly driver (chosen via feature flag) supports hot-swapping module bytes.
     pub fn replace_actor(&self, new_actor: Actor) -> Result<()> {
-        crate::inthost::replace_actor(&self.key, self.bus.clone(), new_actor)
+        let key = KeyPair::from_seed(&self.sk).unwrap();
+        crate::inthost::replace_actor(&key, self.bus.clone(), new_actor)
     }
 
     /// Adds a middleware item to the middleware processing pipeline
@@ -519,6 +481,7 @@ impl Host {
             capability.descriptor().clone(),
         );
         let wg = crossbeam_utils::sync::WaitGroup::new();
+        let key = KeyPair::from_seed(&self.sk).unwrap();
         spawns::spawn_native_capability(
             capability,
             self.bus.clone(),
@@ -527,10 +490,33 @@ impl Host {
             self.terminators.clone(),
             self.plugins.clone(),
             wg.clone(),
-            Arc::new(self.key.clone()),
+            Arc::new(key),
         )?;
         wg.wait();
         Ok(())
+    }
+
+    /// Adds a native capability provider plugin to the host runtime by pulling the library from a provider archive
+    /// stored in an OCI-compliant registry. This file will be stored in the operating system's designated temporary
+    /// directory after being downloaded.
+    pub fn add_native_capability_from_registry(
+        &self,
+        image_ref: &str,
+        binding_name: Option<String>,
+    ) -> Result<()> {
+        let b = binding_name.unwrap_or("default".to_string());
+        match crate::inthost::fetch_provider(image_ref, &b, self.labels.clone()) {
+            Ok((prov, claims)) => {
+                self.add_native_capability(prov)?;
+                // Only write to the image map if the above add function succeeds
+                self.image_map
+                    .write()
+                    .unwrap()
+                    .insert(image_ref.to_string(), claims.subject.to_string());
+                Ok(())
+            }
+            Err(e) => Err(e),
+        }
     }
 
     /// Removes a native capability provider plugin from the waSCC runtime
@@ -568,9 +554,10 @@ impl Host {
         };
         let buf = serialize(&cfg).unwrap();
         let binding = binding_name.unwrap_or("default".to_string());
+        let key = KeyPair::from_seed(&self.sk).unwrap();
         let inv_r = self.bus.invoke(
             &self.bus.provider_subject(&capid, &binding), // The OP_REMOVE_ACTOR invocation should go to _all_ instances of the provider being unbound
-            crate::inthost::gen_remove_actor(&self.key, buf.clone(), &binding, &capid),
+            crate::inthost::gen_remove_actor(&key, buf.clone(), &binding, &capid),
         )?;
         if let Some(s) = inv_r.error {
             Err(format!("Failed to remove binding: {}", s).into())
@@ -597,6 +584,8 @@ impl Host {
         let claims = self.bus.discover_claims(actor);
         #[cfg(not(feature = "lattice"))]
         let claims = self.claims.read().unwrap().get(actor).cloned();
+
+        let key = KeyPair::from_seed(&self.sk).unwrap();
 
         if claims.is_none() {
             return Err(errors::new(errors::ErrorKind::MiscHost(
@@ -639,7 +628,7 @@ impl Host {
         };
         trace!("Binding subject: {}", tgt_subject);
         let inv = inthost::gen_config_invocation(
-            &self.key,
+            &key,
             actor,
             capid,
             c.clone(),
@@ -686,13 +675,14 @@ impl Host {
     /// make a lattice-wide call. If you want to make lattice-wide invocations, please use
     /// the lattice client library.
     pub fn call_actor(&self, actor: &str, operation: &str, msg: &[u8]) -> Result<Vec<u8>> {
+        let key = KeyPair::from_seed(&self.sk).unwrap();
         if !self.claims.read().unwrap().contains_key(actor) {
             return Err(errors::new(errors::ErrorKind::MiscHost(
                 "No such actor".into(),
             )));
         }
         let inv = Invocation::new(
-            &self.key,
+            &key,
             WasccEntity::Actor(SYSTEM_ACTOR.to_string()),
             WasccEntity::Actor(actor.to_string()),
             operation,
@@ -729,15 +719,18 @@ impl Host {
             }
         }
         for actor in manifest.actors {
-            #[cfg(feature = "lattice")]
-            self.add_actor_gantry_first(&actor)?;
-
-            #[cfg(not(feature = "lattice"))]
-            self.add_actor(Actor::from_file(&actor)?)?;
+            self.add_actor_file_first(&actor)?; // If file, add .wasm, otherwise assume it's an OCI ref
         }
         for cap in manifest.capabilities {
             // for now, supports only file paths
-            self.add_native_capability(NativeCapability::from_file(cap.path, cap.binding_name)?)?;
+            if Path::new(&cap.path).exists() {
+                self.add_native_capability(NativeCapability::from_file(
+                    cap.path,
+                    cap.binding_name,
+                )?)?;
+            } else {
+                self.add_native_capability_from_registry(&cap.path, cap.binding_name)?;
+            }
         }
         for config in manifest.bindings {
             self.set_binding(
@@ -750,13 +743,11 @@ impl Host {
         Ok(())
     }
 
-    #[cfg(feature = "lattice")]
-    fn add_actor_gantry_first(&self, actor: &str) -> Result<()> {
-        if actor.len() == 56 && actor.starts_with('M') {
-            // This is an actor's public subject
-            self.add_actor_from_gantry(actor, 0) // TODO: do not default to the latest revision when using manifest file
-        } else {
+    fn add_actor_file_first(&self, actor: &str) -> Result<()> {
+        if std::path::Path::new(actor).exists() {
             self.add_actor(Actor::from_file(&actor)?)
+        } else {
+            self.add_actor_from_registry(actor)
         }
     }
 
@@ -816,6 +807,6 @@ impl Host {
 
     /// Returns the public key of the host
     pub fn id(&self) -> String {
-        self.key.public_key()
+        self.pk.to_string()
     }
 }