+++ src/spawns.rs
@@ -44,13 +44,18 @@ pub(crate) fn spawn_actor(
     terminators: Arc<RwLock<HashMap<String, Sender<bool>>>>,
     hk: KeyPair,
     auth: Arc<RwLock<Box<dyn Authorizer>>>,
+    image_map: Arc<RwLock<HashMap<String, String>>>,
+    imgref: Option<String>,
 ) -> Result<()> {
     let c = claims.clone();
     let b = bus.clone();
-    let hostkey = hk.clone();
+    let seed = hk.seed().unwrap();
+    let s = seed.clone();
+    let hostkey = KeyPair::from_seed(&hk.seed().unwrap()).unwrap();
     let authorizer = auth.clone();
 
     thread::spawn(move || {
+        let hk = KeyPair::from_seed(&seed).unwrap();
         if actor {
             #[cfg(feature = "lattice")]
             let _ = bus.publish_event(BusEvent::ActorStarting {
@@ -64,8 +69,9 @@ pub(crate) fn spawn_actor(
         let engine = wasm3_provider::Wasm3EngineProvider::new(&buf);
 
         let mut guest = WapcHost::new(Box::new(engine), move |_id, bd, ns, op, payload| {
+            let key = KeyPair::from_seed(&s).unwrap();
             wapc_host_callback(
-                hk.clone(),
+                key,
                 c.clone(),
                 bus.clone(),
                 bd,
@@ -123,6 +129,7 @@ pub(crate) fn spawn_actor(
             info!("Actor {} up and running.", &claims.subject);
         }
         loop {
+            let key = KeyPair::from_seed(&seed).unwrap();
             select! {
                 recv(inv_r) -> inv => {
                     if let Ok(inv) = inv {
@@ -153,10 +160,16 @@ pub(crate) fn spawn_actor(
                     } else {
                         #[cfg(feature = "lattice")]
                         let _ = b.publish_event(BusEvent::ActorStopped{ host: hostkey.public_key(), actor: claims.subject.to_string() });
+
                         let mut lock = claimsmap.write().unwrap();
                         let _ = lock.remove(&claims.subject);
                         drop(lock);
-                        deconfigure_actor(hostkey.clone(),b.clone(), bindings.clone(), &claims.subject);
+                        if let Some(ref ir) = imgref { // if this actor was added via OCI image ref, remove the mapping
+                            let mut lock = image_map.write().unwrap();
+                            let _ = lock.remove(ir);
+                            drop(lock);
+                        }
+                        deconfigure_actor(key,b.clone(), bindings.clone(), &claims.subject);
                     }
                     break "".to_string(); // TODO: WHY WHY WHY does this recv arm need to return a value?!?!?
                 }